[DIDA] feat. 게시글 작성 > Empty View 추가

+++ feature/create-community-input/src/main/java/com/dida/create_community_input/CreateCommunityInputViewModel.kt
@@ -75,7 +75,7 @@ class CreateCommunityInputViewModel @AssistedInject constructor(
             postIdState.value = postId
             postIdAPI(postId = postId)
                 .onSuccess {
-                    _cardIdState.value = it.cardId.toLong()
+                    _cardIdState.value = it.cardId
                     _cardImgState.value = it.cardImgUrl
                     _cardTitleState.value = it.title
                     _userImgState.value = it.userImgUrl
